taxmate_backend/requirements.txt
django>=4.2.0
djangorestframework>=3.14.0
django-cors-headers>=4.1.0
mysqlclient>=2.1.1
nltk>=3.9.1
scikit-learn>=1.6.1
gensim>=4.3.0